Avoided a few more warnings with GCC -Wconversion
git-svn-id: svn://svn.code.sf.net/p/fuego/code/trunk@1186 2e953b5c-c64d-0410-be54-f773e93e544c

+++ smartgame/SgHash.h
@@ -183,7 +183,7 @@ template<int N>
 unsigned int SgHash<N>::GetWord() const
 {
     static const std::bitset<N> mask(0xffffffffUL);
-    return (m_code & mask).to_ulong();
+    return (unsigned int)((m_code & mask).to_ulong());
 }
 
 template<int N>
